backend/requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
colorclass==2.2.2
Django==4.2
djangorestframework==3.14.0
docopt==0.6.2
idna==3.4
packaging==23.0
pip-upgrader==1.4.15
python-dotenv==1.0.0
pytz==2023.3
requests==2.28.2
sqlparse==0.4.3
terminaltables==3.1.10
urllib3==1.26.15
